# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SystemStatus = typeof SystemStatus[keyof typeof SystemStatus];


export const SystemStatus = {
  online: 'online',
  warning: 'warning',
  offline: 'offline',
} as const;

export type SystemDerivedStatus = typeof SystemDerivedStatus[keyof typeof SystemDerivedStatus];


export const SystemDerivedStatus = {
  online: 'online',
  warning: 'warning',
  offline: 'offline',
} as const;

export type SystemMetadata = { [key: string]: unknown };

export interface System {
  id: string;
  name: string;
  status: SystemStatus;
  derived_status: SystemDerivedStatus;
  last_ping: string;
  metadata: SystemMetadata;
  created_at: string;
}

export type SystemLogPayload = { [key: string]: unknown };

export interface SystemLog {
  id: string;
  system_name: string;
  event_type: string;
  /** @nullable */
  message: string | null;
  payload: SystemLogPayload;
  created_at: string;
}

export interface DashboardSummary {
  total_systems: number;
  online_systems: number;
  warning_systems: number;
  offline_systems: number;
  events_last_24h: number;
  latest_event: SystemLog | null;
}

export type TelemetryInputStatus = typeof TelemetryInputStatus[keyof typeof TelemetryInputStatus];


export const TelemetryInputStatus = {
  online: 'online',
  warning: 'warning',
  offline: 'offline',
} as const;

export type TelemetryInputPayload = { [key: string]: unknown };

export interface TelemetryInput {
  /** @minLength 1 */
  system_name: string;
  status: TelemetryInputStatus;
  /** @minLength 1 */
  event_type: string;
  message?: string;
  payload?: TelemetryInputPayload;
}

export interface TelemetryResponse {
  success: boolean;
  system: System;
  log: SystemLog;
}

export type ErrorResponseDetails = { [key: string]: unknown };

export interface ErrorResponse {
  error: string;
  details?: ErrorResponseDetails;
}

export type TelegramSendInputPriority = typeof TelegramSendInputPriority[keyof typeof TelegramSendInputPriority];


export const TelegramSendInputPriority = {
  normal: 'normal',
  urgent: 'urgent',
} as const;

export type TelegramSendInputChannelsItem = typeof TelegramSendInputChannelsItem[keyof typeof TelegramSendInputChannelsItem];


export const TelegramSendInputChannelsItem = {
  telegram: 'telegram',
  discord: 'discord',
} as const;

export interface TelegramSendInput {
  /**
     * @minLength 1
     * @maxLength 4096
     */
  message: string;
  priority?: TelegramSendInputPriority;
  /** @minItems 1 */
  channels?: TelegramSendInputChannelsItem[];
}

export type TelegramSendResponsePriority = typeof TelegramSendResponsePriority[keyof typeof TelegramSendResponsePriority];


export const TelegramSendResponsePriority = {
  normal: 'normal',
  urgent: 'urgent',
} as const;

export interface TelegramSendResponse {
  success: boolean;
  message: string;
  priority: TelegramSendResponsePriority;
  /** @nullable */
  telegram_message_id: number | null;
  discord_delivered: boolean;
  log_id: string;
}

export interface NotificationStatus {
  telegram_configured: boolean;
  discord_configured: boolean;
}

export type PipelineTriggerInputSource = typeof PipelineTriggerInputSource[keyof typeof PipelineTriggerInputSource];


export const PipelineTriggerInputSource = {
  dashboard: 'dashboard',
  cron: 'cron',
} as const;

export interface PipelineTriggerInput {
  source?: PipelineTriggerInputSource;
}

export type PipelineTriggerResponseStatus = typeof PipelineTriggerResponseStatus[keyof typeof PipelineTriggerResponseStatus];


export const PipelineTriggerResponseStatus = {
  queued: 'queued',
  running: 'running',
} as const;

export interface PipelineTriggerResponse {
  accepted: boolean;
  run_id: string;
  status: PipelineTriggerResponseStatus;
  message: string;
  log_id: string;
}

export type ApiTesterInputMethod = typeof ApiTesterInputMethod[keyof typeof ApiTesterInputMethod];


export const ApiTesterInputMethod = {
  GET: 'GET',
  POST: 'POST',
  PUT: 'PUT',
  PATCH: 'PATCH',
  DELETE: 'DELETE',
} as const;

export type ApiTesterInputHeaders = {[key: string]: string};

export interface ApiTesterInput {
  /** @maxLength 4096 */
  url: string;
  method: ApiTesterInputMethod;
  headers?: ApiTesterInputHeaders;
  /** @nullable */
  body?: string | null;
}

export type ApiTesterResponseRequest = {
  url: string;
  method: string;
};

export type ApiTesterResponseHeaders = {[key: string]: string};

export interface ApiTesterResponse {
  request: ApiTesterResponseRequest;
  status: number;
  status_text: string;
  headers: ApiTesterResponseHeaders;
  body: unknown;
  body_text: string;
  elapsed_ms: number;
}

export interface MediaExtractInput {
  /** @maxLength 4096 */
  url: string;
}

export type MediaExtractResponseStatus = typeof MediaExtractResponseStatus[keyof typeof MediaExtractResponseStatus];


export const MediaExtractResponseStatus = {
  accepted: 'accepted',
} as const;

export type MediaExtractResponsePlatform = typeof MediaExtractResponsePlatform[keyof typeof MediaExtractResponsePlatform];


export const MediaExtractResponsePlatform = {
  tiktok: 'tiktok',
  instagram: 'instagram',
  reddit: 'reddit',
  x: 'x',
} as const;

export interface MediaExtractResponse {
  status: MediaExtractResponseStatus;
  message: string;
  url: string;
  platform: MediaExtractResponsePlatform;
  /** @nullable */
  job_id: string | null;
}

export type ListSystemLogsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
system_name?: string;
};