# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /api-tester/execute:
    post:
      operationId: executeApiTesterRequest
      tags: [api-tester]
      summary: Execute an HTTP request
      description: Executes a validated HTTP request from the API server and returns the upstream response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiTesterInput"
      responses:
        "200":
          description: Upstream response received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiTesterResponse"
        "400":
          description: Invalid request input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Upstream response exceeded the size limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Upstream request failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /system-logs:
    get:
      operationId: listSystemLogs
      tags: [dashboard]
      summary: List recent system logs
      description: Returns the latest telemetry activity, newest first.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 100
            default: 30
        - name: system_name
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Recent system logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SystemLog"
...
  /telemetry:
    post:
      operationId: ingestTelemetry
      tags: [telemetry]
      summary: Ingest worker telemetry
      description: Validates and records a telemetry event, then updates the related system.
      parameters:
        - name: X-API-KEY
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelemetryInput"
      responses:
        "200":
          description: Telemetry accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetryResponse"
        "400":
          description: Invalid telemetry body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid telemetry API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /pipeline/trigger:
    post:
      operationId: triggerPipeline
      tags: [pipeline]
      summary: Trigger the daily pipeline
      description: Starts the shared Scout, Council, and Dispatch runner when the pipeline is configured.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PipelineTriggerInput"
      responses:
        "202":
          description: Pipeline accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineTriggerResponse"
        "401":
          description: Trigger authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Another pipeline run is in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Manual trigger rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Pipeline is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    System:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [online, warning, offline]
        derived_status:
          type: string
          enum: [online, warning, offline]
        last_ping:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
      required: [id, name, status, derived_status, last_ping, metadata, created_at]
    SystemLog:
      type: object
      properties:
        id:
          type: string
        system_name:
          type: string
        event_type:
          type: string
        message:
          type: [string, "null"]
        payload:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
      required: [id, system_name, event_type, message, payload, created_at]
...
    TelemetryInput:
      type: object
      properties:
        system_name:
          type: string
          minLength: 1
        status:
          type: string
          enum: [online, warning, offline]
        event_type:
          type: string
          minLength: 1
        message:
          type: string
        payload:
          type: object
          additionalProperties: true
      required: [system_name, status, event_type]
    TelemetryResponse:
      type: object
      properties:
        success:
          type: boolean
        system:
          $ref: "#/components/schemas/System"
        log:
          $ref: "#/components/schemas/SystemLog"
      required: [success, system, log]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [error]
...
    PipelineTriggerInput:
      type: object
      properties:
        source:
          type: string
          enum: [dashboard, cron]
          default: dashboard
    PipelineTriggerResponse:
      type: object
      properties:
        accepted:
          type: boolean
        run_id:
          type: string
        status:
          type: string
          enum: [queued, running]
        message:
          type: string
        log_id:
          type: string
      required: [accepted, run_id, status, message, log_id]
    ApiTesterInput:
      type: object
      properties:
        url:
          type: string
          maxLength: 4096
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
        headers:
          type: object
          additionalProperties:
            type: string
          default: {}
        body:
          type: ["string", "null"]
      required: [url, method]
    ApiTesterResponse:
      type: object
      properties:
        request:
          type: object
          properties:
            url:
              type: string
            method:
              type: string
          required: [url, method]
        status:
          type: number
        status_text:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          {}
        body_text:
          type: string
        elapsed_ms:
          type: number
      required: [request, status, status_text, headers, body, body_text, elapsed_ms]
...