# lib/api-zod/src/generated/types/apiTesterResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApiTesterResponseHeaders } from './apiTesterResponseHeaders';
import type { ApiTesterResponseRequest } from './apiTesterResponseRequest';

export interface ApiTesterResponse {
  request: ApiTesterResponseRequest;
  status: number;
  status_text: string;
  headers: ApiTesterResponseHeaders;
  body: unknown;
  body_text: string;
  elapsed_ms: number;
}