# lib/api-zod/src/generated/types/telemetryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TelemetryInputPayload } from './telemetryInputPayload';
import type { TelemetryInputStatus } from './telemetryInputStatus';

export interface TelemetryInput {
  /** @minLength 1 */
  system_name: string;
  status: TelemetryInputStatus;
  /** @minLength 1 */
  event_type: string;
  message?: string;
  payload?: TelemetryInputPayload;
}