# lib/api-zod/src/generated/types/telemetryResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { System } from './system';
import type { SystemLog } from './systemLog';

export interface TelemetryResponse {
  success: boolean;
  system: System;
  log: SystemLog;
}